+++ R/CFObject.R
@@ -15,6 +15,9 @@ CFObject <- R6::R6Class("CFObject",
     #' @field NCvar The [NCVariable] instance that this CF object represents.
     NCvar = NULL,
 
+    #' @field group The [NCGroup] that this object is located in.
+    group = NULL,
+
     #' @description Create a new CF object instance from a variable in a netCDF
     #'   resource. This method is called upon opening a netCDF resource. It is
     #'   rarely, if ever, useful to call this constructor directly from the
@@ -23,9 +26,11 @@ CFObject <- R6::R6Class("CFObject",
     #'
     #' @param nc_var The [NCVariable] instance upon which this CF object is
     #'   based.
+    #' @param group The [NCGroup] that this object is located in.
     #' @return A `CFobject` instance.
-    initialize = function(nc_var) {
+    initialize = function(nc_var, group) {
       self$NCvar <- nc_var
+      self$group <- group
     },
 
     #' @description Retrieve attributes of any CF object.
@@ -74,6 +79,15 @@ CFObject <- R6::R6Class("CFObject",
     name = function(value) {
       if (missing(value))
         self$NCvar$name
+    },
+
+    #' @field fullname (read-only) The fully-qualified name of the CF object.
+    fullname = function(value) {
+      if (missing(value)) {
+        grp <- self$group$fullname
+        if (!(grp == "/")) grp <- paste0(grp, "/")
+        paste0(grp, self$NCvar$name)
+      }
     },
 
     #' @field attributes (read-only) A `data.frame` with the attributes of the